% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter.R
\name{apply_filters}
\alias{apply_filters}
\title{Filter gene counts matrix by applying various filters}
\usage{
apply_filters(
  count_matrix,
  median_min = 1,
  expression_sum_min = 10,
  max_min = 5,
  range_min = 5,
  prop_zero_max = 0.8,
  cv_min = NULL,
  cv_max = NULL,
  max_to_median_max = NULL
)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}

\item{median_min}{minimum allowed median count value in the sample. Default is 1.}

\item{expression_sum_min}{minimum allowed sum of counts across individuals. Default is 10.}

\item{max_min}{minimum allowed maximum value in the sample for a gene. Default is 5.}

\item{range_min}{minimum allowed range of counts in the sample. Default is 5.}

\item{prop_zero_max}{maximum allowed proportion of individuals with count of
zero for a transcript. Default is 0.8.}

\item{cv_min}{minimum allowed value of coefficient of variation. Default is NULL.}

\item{cv_max}{maximum allowed value of coefficient of variation. Default is NULL.}

\item{max_to_median_max}{maximum allowed ratio between the maximum to the}
}
\value{
Matrix of gene counts, with filtered (reduced) rows.
}
\description{
The purpose of filtering is to remove lowly express genes
}
\examples{
data(rnaseq_count_matrix)
apply_filters(count_matrix=rnaseq_count_matrix, median_min = 1, expression_sum_min = 10,max_min = 5,
              range_min = 5, prop_zero_max = 0.8,cv_min = NULL, cv_max = NULL,
              max_to_median_max = NULL)

}
