% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_empirical_pvalue.R
\name{compute_empirical_pvalues}
\alias{compute_empirical_pvalues}
\title{Computing empirical pvalue}
\usage{
compute_empirical_pvalues(
  statistics,
  null_statistics,
  empirical_type = "storey",
  stat_type = "t_stat",
  t_df = NULL
)
}
\arguments{
\item{statistics}{A vector of calculated test statistics}

\item{null_statistics}{A vactor of permuted test statistic}

\item{empirical_type}{Type of empirical pvalue : quantile or storey}

\item{stat_type}{Statistic type : p_value (quantile empirical pvalue), t_stat and chisq_stat (storey)}

\item{t_df}{A vector of calculated t-statistic}
}
\value{
A vector of empirical p-value
}
\description{
Computing empirical pvalue
}
\examples{
library(qvalue)
stat<- runif(1000,0,1)
null_stat<- runif(100000,0,1)
compute_empirical_pvalues(statistics=stat,null_statistics=null_stat,
                         empirical_type = "quantile",stat_type = "p_value",
                         t_df = NULL)
}
