% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_empirical_pvalue.R
\name{compute_quantile_empirical_pvalues}
\alias{compute_quantile_empirical_pvalues}
\title{Quantile empirical P-values}
\usage{
compute_quantile_empirical_pvalues(statistics, null_statistics)
}
\arguments{
\item{statistics}{A vector of calculated p-value.}

\item{null_statistics}{A vactor of permuted test pvalue}
}
\value{
A vector of empirical pvalues
}
\description{
Quantile empirical P-values
}
\examples{
stat<- runif(1000,0,1)
null_stat<- runif(100000,0,1)
compute_quantile_empirical_pvalues(statistics=stat,null_statistics=null_stat)
}
