% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_empirical_pvalue.R
\name{compute_storey_empirical_pvalues}
\alias{compute_storey_empirical_pvalues}
\title{Storey empirical P-values}
\usage{
compute_storey_empirical_pvalues(statistics, null_statistics)
}
\arguments{
\item{statistics}{A vector of calculated test statistics}

\item{null_statistics}{A vactor of permuted test statistic}
}
\value{
A vector of empirical pvalues
}
\description{
Storey empirical P-values
}
\examples{
library(qvalue)
stat<- rnorm(1000,0,1)
null_stat<- rnorm(100000,0,1)
compute_storey_empirical_pvalues(statistics=stat,null_statistics=null_stat)

}
\references{
Storey J, Bass A, Dabney A, Robinson D. 2019. qvalue: Q-value estimation for false discovery rate control. In R package version 2.18.0
}
