% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{ENSG00000000003}
\alias{ENSG00000000003}
\title{Single gene count (ENSG00000000003)}
\format{
a vector which contains  40 column and match with rownames in phenotypes
\describe{
  \item{column}{sample names and it must be matched with the rownames in the phenotype}
 @source \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE151243}
}
}
\usage{
ENSG00000000003
}
\description{
Example of single gene count
}
\keyword{datasets}
