% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_genes.R
\name{filter_by_genes}
\alias{filter_by_genes}
\title{Filter gene counts matrix by geneID}
\usage{
filter_by_genes(count_matrix, gene_IDs)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}

\item{gene_IDs}{A vactor of gene names}
}
\value{
Matrix of selected gene counts
}
\description{
Filtering gene counts matrix based on intrested gene/s
}
\examples{
data(rnaseq_count_matrix)
genes<-c("ENSG00000000003","ENSG00000000005","ENSG00000000419")
filter_by_genes(count_matrix=rnaseq_count_matrix, gene_IDs=genes)
}
