% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_count_mat.R
\name{lm_count_mat}
\alias{lm_count_mat}
\title{Fast linear regression}
\usage{
lm_count_mat(
  count_matrix,
  pheno,
  trait,
  covariates_string,
  gene_IDs = NULL,
  log_transform = "log_replace_half_min"
)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}

\item{pheno}{A data frame of phenotype, includes the trait and covariates.}

\item{trait}{A character, the name of the exposure variable. The trait should be a column in pheno.}

\item{covariates_string}{Characters string with specifying the covariats, include "as.factor" statements. example: covariate_string = "age,as.factor(sex)"}

\item{gene_IDs}{Vector of selection of geneID, NULL if all genes are tested}

\item{log_transform}{One of the transformations log_replace_half_min, log_add_min, log_add_0.5, or NULL (default)}
}
\value{
Linear regression results as a data frame with columns geneID, beta,se,t_stat (t-statistic),t_stat_df(degree of freedom),p_value,fdr_bh
}
\description{
Calculated associtation gene counts matrix with the pehonotype where each of gene counts on aexposure
}
\examples{
library(dplyr)
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
data(phenotype)
trait<-"Trait.1"
covars<-"Age+Sex"
log_transform<-"log_replace_half_min"
lm_count_mat(count_matrix=rnaseq_count_matrix,pheno=phenotype,trait=trait,
             covariates_string=covars, log_transform=log_transform)
}
