% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_lm_count_mat.R
\name{lm_mult_count_mat_emp_pval}
\alias{lm_mult_count_mat_emp_pval}
\title{Wrapper function for differential expression analysis for multiple exposure}
\usage{
lm_mult_count_mat_emp_pval(
  count_matrix,
  pheno,
  traits,
  covariates_string,
  n_permute = 100,
  gene_IDs = NULL,
  log_transform = "log_replace_half_min",
  seed = NULL,
  stat_type = "z_score",
  empirical_type = "storey",
  family = "gaussian",
  t_df = NULL
)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}

\item{pheno}{A data frame phenotype data, includes the trait and covariates.}

\item{traits}{Characters, the name of the exposure variable. The traits should be a column in pheno.}

\item{covariates_string}{A character string with specifying the covariats, include "as.factor" statements. example: covariate_string = "age,as.factor(sex)"}

\item{n_permute}{number of computing residual permutation. Default is 100 times}

\item{gene_IDs}{: vector of selection of geneID, NULL if all genes are tested}

\item{log_transform}{One of the transformations log_replace_half_min, log_add_min, log_add_0.5, or NULL (default)}

\item{seed}{Random seed}

\item{stat_type}{Statistic type : p_value (quantile empirical pvalue), t_stat and z_score (storey). Default is z_score}

\item{empirical_type}{Type of empirical pvalue : quantile or storey. Default  is storey}

\item{family}{A description of the error distribution to be used in the model. gaussian if the variable is continous,
and binomial if variable is binary. The default is "gaussian"}

\item{t_df}{A vector of calculated t-statistic.Default is NULL}
}
\value{
Linear regression results as a data frame with columns geneID, beta.Trait1,beta.Trait1 ,se,
       t_stat (join t-statistic),p_value(join p-value),z_score (transformed by p-value),emp_pvals,bh_emp_pvals
}
\description{
The function is to calculate DEG (Differential Expression Genes) analysis for multiple exposure using residual permuation approach to calculate empirical p-value
}
\examples{
set.seed(123)
library(qvalue)
library(dplyr)
data(phenotype)
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
traits<-c("Trait.1","Trait.2")
covars<- "Age,Sex"
lm_mult_count_mat_emp_pval(count_matrix=rnaseq_count_matrix, pheno=phenotype, traits=traits, covariates_string=covars,
                        stat_type="z_score", empirical_type = "storey",family="gaussian")
}
