% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_replace_half_min}
\alias{log_replace_half_min}
\title{Log replace half min}
\usage{
log_replace_half_min(count_matrix)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}
}
\value{
A matrix of gene counts after transformation
}
\description{
Type of log transform, replacing zero with minimum value for a gene divided by 2
}
\examples{
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
log_replace_half_min(count_matrix=rnaseq_count_matrix)
}
