% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform_count}
\alias{log_transform_count}
\title{Applying log transform on a matrix of gene counts}
\usage{
log_transform_count(count_matrix, transform = "log_replace_half_min")
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}

\item{transform}{One of the transformations log_replace_half_min, log_add_min, log_add_0.5}
}
\value{
A matrix of gene counts after transformation
}
\description{
Applying log transform on a matrix of gene counts
}
\examples{
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
log_transform_count(count_matrix=rnaseq_count_matrix, transform = "log_replace_half_min")
}
