% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{median_normalization}
\alias{median_normalization}
\title{Median Normalization}
\usage{
median_normalization(count_matrix)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}
}
\value{
median_norm  A matrix of gene counts  after normalization
}
\description{
Median normalization is using median value for each library so they have equal size
}
\examples{
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
median_normalization(rnaseq_count_matrix)
}
