% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_lm_count_mat.R
\name{mult_lm_count_mat}
\alias{mult_lm_count_mat}
\title{Fast linear regression for multiple exposure}
\usage{
mult_lm_count_mat(
  count_matrix,
  pheno,
  covariates_string,
  traits,
  gene_IDs = NULL,
  log_transform = "log_replace_half_min"
)
}
\arguments{
\item{count_matrix}{A matrix of gene counts (possibly transformed). rows are genes, columns are individuals}

\item{pheno}{A data frame of phenotype data, includes the trait and covariates.}

\item{covariates_string}{A character string with specifying the covariates, include "as.factor" statements. example: covariates_string = "age + as.factor(sex)"}

\item{traits}{Characters, the name of the exposure variables. The traits should columns in pheno.}

\item{gene_IDs}{A vector of selection of gene IDs, NULL if all genes are tested}

\item{log_transform}{One of the transformations log_replace_half_min, log_add_min, log_add_0.5, or NULL (default)}
}
\value{
Linear regression results as a data frame with columns geneID, beta.Trait.1,beta.Trait.1 ,se,t_stat (join t-statistic),p_value(join p-value),fdr_bh ,z_score (transformed by p-value)
}
\description{
Calculated associtation gene counts matrix with the pehonotype where each of gene counts on multiple traits
}
\examples{
set.seed(123)
library(dplyr)
data(phenotype)
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
traits<-c("Trait.1","Trait.2")
covars<- "Age,Sex"
mult_lm_count_mat(count_matrix=rnaseq_count_matrix,pheno = phenotype,traits = traits,
                  covariates_string=covars)
}
