% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_permutation_function.R
\name{permute_resids_trait_cor}
\alias{permute_resids_trait_cor}
\title{Power Simulations}
\usage{
permute_resids_trait_cor(
  pheno,
  trait,
  covariates_string,
  gene_exp,
  required_cor,
  seed = NULL,
  family = "gaussian"
)
}
\arguments{
\item{pheno}{A data.frame with phenotypes}

\item{trait}{A continuous/ binary variable to permute while preserving association with covariates}

\item{covariates_string}{Covariates to adjust for, as a string used in regression model}

\item{gene_exp}{Selected gene/transcript to simulate as associated with the trait}

\item{required_cor}{Corelations value simulating the strength of association between trait and gene_exp}

\item{seed}{Random seed}

\item{family}{A description of the error distribution to be used in the model. gaussian if the variable is continous,
and binomial if variable is binary. The default is "gaussian"}
}
\value{
A vector permuted trait that is associated with gene_exp
}
\description{
Power Simulations
}
\examples{
data(phenotype)
data(ENSG00000000003)
trait<-"Trait.1"; covars="Age,Sex"
permute_resids_trait_cor(pheno= phenotype,trait=trait,
                         covariates_string=covars, required_cor=0.3,
                         gene_exp=ENSG00000000003,seed=NULL, family="gaussian")
}
