% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{Simulated phenotypes.}
\format{
A data frame with 40 rows and 4 variables, and row names must be matched with column names in rnaseq_count_matrix:
\describe{
  \item{Age}{covariate, must be integer}
  \item{Sex}{covariate,0 is female and 1 is male, binary}
  \item{Trait.1}{outcome/ trait , numeric}
  \item{Trait.2}{outcome/ trait , numeric}
  ...
}
}
\usage{
phenotype
}
\description{
A dataset containing simultaed 40 samples
}
\keyword{datasets}
