% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{residual_plot}
\alias{residual_plot}
\title{Title Residual plot}
\usage{
residual_plot(pheno, covariates_string, traits)
}
\arguments{
\item{pheno}{A data frame of phenotype, includes the trait and covariates}

\item{covariates_string}{Characters string with specifying the covariats, include "as.factor" statements. example: covariate_string = "age,as.factor(sex)"}

\item{traits}{Trait/s, the name of the exposure variable. The trait/s should be a column in pheno.}
}
\value{
residual plot
}
\description{
Title Residual plot
}
\examples{
library(ggplot2)
library(reshape2)
data(phenotype)
covariates_string<-"Age+Sex+Race"
traits<-c("Trait.1","Trait.2")
residual_plot(pheno=phenotype, covariates_string = covariates_string,traits=traits)
}
