% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{rnaseq_count_matrix}
\alias{rnaseq_count_matrix}
\title{Raw gene count matrix}
\format{
a matrix which contains 58051 and 40 column.
\describe{
  \item{column}{sample names and it must be matched with the rownames in the phenotype}
  \item{rownames}{ensemblID/ geneID }
 @source \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE151243}
}
}
\usage{
rnaseq_count_matrix
}
\description{
Example of bulk RNA-seq
}
\keyword{datasets}
