% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{summarize_phenotypes}
\alias{summarize_phenotypes}
\title{Title Summarize phenotypes}
\usage{
summarize_phenotypes(pheno, categorical_variables, numeric_variables, strata)
}
\arguments{
\item{pheno}{A data frame of phenotypes, includes the trait and covariates}

\item{categorical_variables}{all selected categorical variables in the model, example: c("Sex", "Race")}

\item{numeric_variables}{all selected numeric variables in the model, example: c("Age", "Trait.1", "Trait.2")}

\item{strata}{Variable to stratify, to example: "Race"}
}
\value{
summary table of phenotypes
}
\description{
Title Summarize phenotypes
}
\examples{
library(tableone)
data(phenotype)
categorical_variable<-c("Sex","Race")
numeric_variable<-c("Age","Trait.1","Trait.2")
strata<-"Race"
summarize_phenotypes(pheno=phenotype, categorical_variables =categorical_variables,  numeric_variables=numeric_variable, strata=strata)
}
