% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{summary_phenotype}
\alias{summary_phenotype}
\title{Title Summary phenotype}
\usage{
summary_phenotype(pheno, categorical_variable, numeric_variable, strata)
}
\arguments{
\item{pheno}{A data frame of phenotype, includes the trait and covariates}

\item{categorical_variable}{all selected categorical variable in model example: c("Sex", "Race")}

\item{numeric_variable}{all selected numeric variable in model example: c("Age", "Trait.1", "Trait.2")}

\item{strata}{Variable to stratified example: "Race"}
}
\value{
sumary phenotype
}
\description{
Title Summary phenotype
}
\examples{
library(tableone)
data(phenotype)
categorical_variable<-c("Sex","Race")
numeric_variable<-c("Age","Trait.1","Trait.2")
strata<-"Race"
summary_phenotype(pheno=phenotype, categorical_variable =categorical_variable,  numeric_variable=numeric_variable, strata=strata)
}
