% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{violin_plot}
\alias{violin_plot}
\title{Title Violin plot}
\usage{
violin_plot(pheno, strata, norm_count_matrix, selected_transcript)
}
\arguments{
\item{pheno}{A data frame of phenotype, includes the trait and covariates}

\item{strata}{Variable to stratified example: "Race"}

\item{norm_count_matrix}{A matrix of gene counts (possibly normalize transformed). rows are genes, columns are individuals}

\item{selected_transcript}{transcript selection example: "ENSG00000002549"}

\item{log_transform}{One of the transformations log_replace_half_min, log_add_min, log_add_0.5, or NULL (default)}
}
\value{
violin plot
}
\description{
Title Violin plot
}
\examples{
strata<-"Race"
data(phenotype)
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
selected_transcript<- "ENSG00000002549"
strata<-"Race"
violin_plot(pheno=phenotype, strata=strata, norm_count_matrix=rnaseq_count_matrix, selected_transcript=selected_transcript )
}
