% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Title Volcano plot}
\usage{
volcano_plot(deg_res, significant_threshold)
}
\arguments{
\item{deg_res}{A data frame of results differential of expression genes}

\item{significant_threshold}{threshold for significant genes based on empirical-padj bh, example: 0.1}
}
\value{
volcano plot
}
\description{
Title Volcano plot
}
\examples{
library(ggrepel)
library(dplyr)
data(rnaseq_count_matrix)
rnaseq_count_matrix<- rnaseq_count_matrix[rowSums(rnaseq_count_matrix)>0,]
data(phenotype)
trait<-"Trait.1"
covars<-"Age+Sex"

median_norm<- median_normalization(rnaseq_count_matrix)
clean_count_matrix <- apply_filters(count_matrix = median_norm, median_min = 1, expression_sum_min = 10, 
                                   max_min = 10, range_min = 5, prop_zero_max = 0.5)
deg_res<-lm_count_mat_emp_pval(count_matrix=clean_count_matrix, pheno=phenotype, trait=trait, covariates_string=covars, 
                             n_permute=100, log_transform = "log_replace_half_min",
                             seed = NULL,outcome_type="continuous")
volcano_plot(deg_res=deg_res,significant_threshold=0.1 )
}
