% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_permutation_function.R
\name{permute_resids_trait}
\alias{permute_resids_trait}
\title{Residual permutation}
\usage{
permute_resids_trait(
  pheno,
  trait,
  covariates_string,
  seed = NULL,
  family = "gaussian"
)
}
\arguments{
\item{pheno}{A data.frame of phenotypes/variables}

\item{trait}{A continuous/binary variable to permute while preserving association with covariates}

\item{covariates_string}{Covariates to adjust for, as a string used in regression model}

\item{seed}{Random seed}

\item{family}{A description of the error distribution to be used in the model. gaussian if the variable is continous,
and binomial if variable is binary. The default is "gaussian"}
}
\value{
Vector of trait values after residual permuted
}
\description{
Residual permutation
}
\examples{
data(phenotype)
trait<-"Trait.1"; covars="Age,Sex"
permute_resids_trait(pheno=phenotype, trait=trait,
                     covariates_string=covars, seed = NULL, family="gaussian")
}
